Refactors navigation flow for bookable products and moves order creation from item type specific flow to the general PurchaseViewController.
